# storage/api-docs/api-documentation.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clockia API v1",
        "description": "Versioned API for business-scoped reservation operations.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application base URL"
        }
    ],
    "paths": {
...
        "/api/v1/businesses/{business}/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "List resources for a business",
                "description": "Returns a paginated resource collection scoped to the business in the URL.",
                "operationId": "listBusinessResources",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Terrace"
                        }
                    },
                    {
                        "name": "filter[is_active]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "filter[resource_type_id]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Allowed values: name, -name, created_at, -created_at",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "name"
                        }
                    },
                    {
                        "name": "page[size]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page[number]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated resources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResourceCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "resources:read"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Create a resource for a business",
                "operationId": "createBusinessResource",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResourceUpsertRequest"
                            },
                            "example": {
                                "name": "Terrace Table 1",
                                "resource_type_id": 1,
                                "capacity": 4,
                                "is_active": true
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "resources:write"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BusinessSummary": {
                "required": [
                    "id",
                    "name",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Restaurante Marea Alta"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "publicDescription": {
                        "type": "string",
                        "example": "A cozy seaside restaurant.",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Ocean Drive",
                        "nullable": true
                    },
                    "website": {
                        "type": "string",
                        "example": "https://example.com",
                        "nullable": true
                    },
                    "cancellationPolicy": {
                        "type": "string",
                        "example": "Free cancellation up to 24h before.",
                        "nullable": true
                    },
                    "minCancellationHours": {
                        "type": "integer",
                        "example": 24,
                        "nullable": true
                    },
                    "allowsModification": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ResourceTypeSummary": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mesa"
                    },
                    "description": {
                        "type": "string",
                        "example": "Recurso físico para atención o servicio.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ServiceSummary": {
                "required": [
                    "id",
                    "name",
                    "duration_minutes",
                    "base_price",
                    "requires_payment",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dinner"
                    },
                    "duration_minutes": {
                        "type": "integer",
                        "example": 120
                    },
                    "base_price": {
                        "type": "string",
                        "example": "35.00"
                    },
                    "requires_payment": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "name": [
                                "The name field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "MessageError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
...
            "Resource": {
                "required": [
                    "id",
                    "name",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "business": {
                        "$ref": "#/components/schemas/BusinessSummary"
                    },
                    "name": {
                        "type": "string",
                        "example": "Terrace Table 1"
                    },
                    "resource_type": {
                        "$ref": "#/components/schemas/ResourceTypeSummary"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "minCapacity": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true,
                        "minimum": 1
                    },
                    "isCombinable": {
                        "type": "boolean",
                        "example": false
                    },
                    "publicNotes": {
                        "type": "string",
                        "example": "Near the window.",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ServiceSummary"
                        }
                    },
                    "availabilities_count": {
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "blocks_count": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "reservations_count": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-08T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-08T12:30:00Z"
                    }
                },
                "type": "object"
            },
            "ResourceUpsertRequest": {
                "required": [
                    "name",
                    "resource_type_id",
                    "is_active"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Terrace Table 1",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "resource_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true,
                        "minimum": 1
                    },
                    "minCapacity": {
                        "type": "integer",
                        "nullable": true,
                        "minimum": 1
                    },
                    "isCombinable": {
                        "type": "boolean"
                    },
                    "publicNotes": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "PaginatedResourceCollection": {
                "required": [
                    "data",
                    "links",
                    "meta"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Resource"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}